complete jwtservice and jwtauthenticationfilter

+++ src/main/java/rastak/train/token/TokenRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
     @Query(value = """
   select t from Token t inner join t.user u
   where u.id = :id and (t.expired = false or t.revoked = false)
   """)
-    List<Token> findAllValidTokenByUser(Integer id);
+    List<Token> findAllValidTokenByUser(Long id);
 
     Optional<Token> findByToken(String token);
 }
